app/requirements.txt
aiogram>=3.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
quart>=0.18.3
hypercorn>=0.14.3
python-dotenv>=1.0.0
pytz>=2023.3
boto3>=1.28.0